// src/app/bridge/bridge-swap/swap-type.ts
export enum SwapTokenType {
  ERC20 = 'ERC20',
  MAINNET = 'MAINNET',
  BEP2 = 'BINANCE',
  BEP20 = 'BEP20',
  BINANCEEXCHANGE = 'BINANCE-EXCHANGE'
}

export enum SwapType {
  ERC20_MAIN = 'ERC20->MAINNET',
  MAIN_ERC20 = 'MAINNET->ERC20',
  BINANCE_MAIN = 'BINANCE->MAINNET',
  MAIN_BINANCE = 'MAINNET->BINANCE',
  ERC20_BINANCE = 'ERC20->BINANCE',
  MAIN_BINANCEEXCHANGE = 'MAINNET->BINANCE-EXCHANGE',
  BEP20_MAIN = 'BEP20->MAINNET',
  MAIN_BEP20 = 'MAINNET->BEP20',
}